requirements.txt
colorama
ipdb
ipykernel
matplotlib
numpy
pandas
nbresult
pygeohash
pytest
seaborn
sklearn
yapf
pillow
scipy
google-cloud-storage
tensorflow==2.8.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-resource-usage==0.6.1
jupyter-server==1.13.1
jupyter-server-mathjax==0.2.3
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.2
jupyterlab-widgets==1.0.2
